+++ src/main/java/blackjack/dto/EarningResultDto.java
@@ -1,4 +1,6 @@
 package blackjack.dto;
 
-public record EarningResultDto(String name, long earningAmount) {
+import java.math.BigDecimal;
+
+public record EarningResultDto(String name, BigDecimal earningAmount) {
 }
